+++ usecases/LDpred2_test/unittest/fun.R
@@ -1,6 +1,8 @@
 library(testthat)
-dirTests <- Sys.getenv('DIR_SCRIPTS')
-source(paste0(dirTests, '/fun.R'))
+dirScripts <- Sys.getenv('DIR_SCRIPTS')
+dirTests <- Sys.getenv('DIR_TESTS')
+source(paste0(dirScripts, '/fun.R'))
+
 context("Test functions")
 dta <- data.frame(a=1:3, b=c(1, 'X', 2))
 test_that("Test numeric counts", {
@@ -15,4 +17,30 @@ test_that("Test various functions", {
   expect_true(hasAllColumns(dta, c('a')))
   expect_false(hasAllColumns(dta, c('a','b', 'c')))
   expect_false(hasAllColumns(dta, c('no')))
+})
+
+test_that("Test if variable is NA (isVarNA)", {
+  expect_true(isVarNA(NA))
+  expect_false(isVarNA(NULL))
+  expect_false(isVarNA(c()))
+  expect_false(isVarNA(F))
+  expect_false(isVarNA('a'))
+  expect_false(isVarNA(1:3))
+  expect_false(isVarNA(data.frame()))
+})
+
+# Test data to use for complementSumstats
+# Joining reference to sumstats should result in 3 successful matches on RSID
+sumstats <- bigreadr::fread2(paste0(dirTests, '/unittest/data/sumstats.txt'))
+reference <- bigreadr::fread2(paste0(dirTests, '/unittest/data/hrc37.txt'))
+test_that("Test appending columns to sumstats", {
+  # Error due to that one of the default columns (CHR) is not available in the reference data
+  expect_error(complementSumstats(sumstats, reference, colRsidSumstats='SNP', colRsidRef='ID'))
+  expect_error(complementSumstats(sumstats, reference, colRsidSumstats='BAD COL', colRsidRef='ID'))
+  expect_error(complementSumstats(sumstats, reference, colRsidSumstats='SNP', colRsidRef='BAD COL'))
+  merged <- complementSumstats(sumstats, reference, colRsidSumstats='SNP', colRsidRef='ID', colsKeepReference=c('#CHROM', 'POS'))
+  expect_s3_class(merged, "data.frame")
+  expect_equal(nrow(merged), 9)
+  expect_equal(sum(!is.na(merged$`#CHROM`)), 3)
+  expect_equal(sum(!is.na(merged$POS)), 3)
 })