+++ CHANGELOG.md
@@ -104,6 +104,7 @@ Main changes since release version [1.0.0](https://github.com/comorment/containe
 
 ### Added
 - add option to append ``usecases/LDpred2/ldpred.R`` score output to an existing file
+- add script ``usecases/LDpred2/complementSumstats.R`` to append chromosome and position to summary statistics
 - add polygenic score output tests for ``usecases/LDpred2/ldpred.R``
 - add ``usecases/LDpred2/imputeGenotypes.R`` for imputing genotypes using R-package bigSNPR
 - add ``usecases/LDpred2/calculateLD.R`` for calculation LD using R-package bigSNPR.